% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa.R
\name{epa}
\alias{epa}
\title{Epanechnicov Kern}
\usage{
epa(x)
}
\arguments{
\item{x}{Reelle Zahl, fuer welche die Epanechnikov-Funktion berechnet wird.}
}
\description{
Die Funktion epa() berechnet die Epanechnikov-Kernfunktion.
}
\author{
Ingo Lange (\emph{ilange@mail.uni-mannheim.de})
}

