% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanli.R
\name{fanli}
\alias{fanli}
\title{Teststatistik Fan und Li}
\usage{
fanli(Y, X1, X2, type = "gaussian", const = 1, bandwidth = "automatic",
  return.bw = F, a = NA, h = NA)
}
\arguments{
\item{Y}{nx1 Datastream, Liste oder Matrix, welche die abhaengigen Variablen enthaelt.}

\item{X1}{nxq1 Datastream oder Matrix, welche die q1 Variablen enthaelt, welche in der Nullhypothese als signifikant angesehen werden.}

\item{X2}{nxq2 Datastream oder Matrix, welche die q2 Variablen enthaelt, welche in der Nullhypothese als insignifikant angesehen werden.}

\item{type}{Der fuer die Kernfunktionen zu verwendende Kern. Es stehen neben dem Gauss-Kern (\emph{="gaussian"}), der Epanechnikov-Kern (\emph{="epanechnikov"}) und der Uniform-Kern (\emph{="uniform"}) zur Verfuegung.}

\item{const}{Konstanten mit der die Bandweiten multipliziert werden sollen. Falls nicht weiter spezifiziert, werden die Bandbreiten nicht veraendert. Der Sinn dieser Konstanten ist die Untersuchung der Sensitivitaet der Bandbreitenwahl.}

\item{bandwidth}{Gibt an, ob die Bandbreite automatisch berechnet werden soll (\emph{"automatic"}). Fuer individuelle Bandbreitenwahl waehle \emph{"manually"}.}

\item{return.bw}{Bei der Standardeinstellung (\emph{FALSE}) wird lediglich der Wert der Teststatistik zurück gegeben.}

\item{a}{Gibt die Bandbreite a an, welche fuer die auf der Variablenmenge X1 basierenden Kernfunktion verwendet werden soll. Wird nur fuer \emph{bandwidth="manually"} verwendet.}

\item{h}{Gibt die Bandbreite h an, welche fuer die auf der Gesamtvariablenmenge basierende Kernfunktion gewaehlt werden soll. Wird nur fuer \emph{bandwidth="manually"} verwendet.}
}
\value{
Gibt den Wert der Teststatistik zurueck.
}
\description{
Die Funktion fanli() berechnet die Teststatistik, welche Fan und Li (1996) in ihrer Arbeit
"Consistent Model Specification Tests: Omitted Variables and Semiparamtric Functional Forms"
vorgestellt haben: Fan Y. & Li, Q (1996): Consistent Model Specification Tests: Omitted Variables and Semiparamtric
 Functional Forms. \emph{Econometrica}, 64 (4), 868-890
}
\author{
Ingo Lange (\emph{ilange@mail.uni-mannheim.de})
}

