% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uni.R
\name{uni}
\alias{uni}
\title{Uniform Kern}
\usage{
uni(x)
}
\arguments{
\item{x}{Reelle Zahl, fuer welche die Uniform-Funktion berechnet wird.}
}
\description{
Die Funktion uni() berechnet die Uniform-Kernfunktion.
}
\author{
Ingo Lange (\emph{ilange@mail.uni-mannheim.de})
}

